requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.3.0
distlib==0.3.7
Django==4.0.10
django-cors-headers==4.3.0
djangorestframework==3.14.0
filelock==3.12.4
idna==3.4
pipenv==2023.10.3
platformdirs==3.11.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.6
virtualenv==20.24.5
